requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
efficientnet-pytorch==0.7.1
filelock==3.9.0
fonttools==4.40.0
fsspec==2023.6.0
huggingface-hub==0.15.1
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.2.1
munch==3.0.0
networkx==3.0
numpy==1.24.1
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.2
Pillow==9.3.0
pretrainedmodels==0.7.4
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.28.1
safetensors==0.3.1
segmentation-models-pytorch==0.3.3
six==1.16.0
sympy==1.11.1
timm==0.9.2
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
tqdm==4.65.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
